-Updated with responses to PR feedback -Added OSX checks for Case Sensitivity tests -Fixed a bug within several Enumerable tests that was causing a test directory to be left over. -Increased code coverage slightly for Directory. -Added tests to ensure whitespace file names are acceptable on Unix machines and marked those tests with an ActiveIssue referencing the newly created #2205
