+++ src/System.IO.FileSystem/tests/PortedCommon/EnumerableUtils.cs
@@ -5,8 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.IO;
-using System.Security;
-using System.Threading;
 using System.Threading.Tasks;
 
 namespace EnumerableTests
@@ -33,10 +31,10 @@ namespace EnumerableTests
         }
 
         // test setup and tear-down
-        public void CreateTestDirs()
+        public void CreateTestDirs(String testFolder)
         {
             String currentDir = Directory.GetCurrentDirectory();
-            testDir = Path.Combine(currentDir, "FSEnumeratorTest", Path.GetRandomFileName());
+            testDir = Path.Combine(currentDir, testFolder, Path.GetRandomFileName());
 
             if (Directory.Exists(testDir))
                 Directory.Delete(testDir, true);
@@ -146,13 +144,14 @@ namespace EnumerableTests
 
         public void DeleteTestDirs()
         {
-            bool deleted = false; int attemptsRemaining = 5;
+            bool deleted = false;
+            int attemptsRemaining = 5;
             while (!deleted && attemptsRemaining > 0)
             {
                 try
                 {
                     if (Directory.Exists(testDir))
-                        Directory.Delete(testDir, true);
+                        FailSafeDirectoryOperations.DeleteDirectory(testDir, true);
                     deleted = true;
                     break;
                 }
